requirements.txt
Flask
pyswisseph
requests
supabase
gunicorn
pytz
timezonefinder
